[NAVYA]ADDED: Invoice Generator should now return the following patr of Invoice :->totalNumberOfRides, TotalFare, AverageFarePerRide

+++ src/main/java/com/BridgeLab/InvoiceService/Ride.java
@@ -1,10 +1,10 @@
 package com.BridgeLab.InvoiceService;
 
-public class Ride {   
-	 public int distance;
+public class Ride {  
+	  public double distance;
 	    public int time;
 
-	    public Ride(int distance, int time) {
+	    public Ride(double distance, int time) {
 	        this.distance = distance;
 	        this.time = time;
 	    }
